Documentacion perfecta del controlador auto!

+++ src/main/java/com/demo/api_consecionaria_autos/service/AutoService.java
@@ -3,9 +3,15 @@ package com.demo.api_consecionaria_autos.service;
 import com.demo.api_consecionaria_autos.model.Auto;
 import com.demo.api_consecionaria_autos.repository.IAutoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatusCode;
+import org.springframework.http.ResponseEntity;
 import org.springframework.stereotype.Service;
+import org.springframework.web.bind.annotation.ResponseStatus;
+import org.springframework.web.client.HttpStatusCodeException;
 
 import java.util.List;
+import java.util.NoSuchElementException;
+import java.util.Optional;
 
 @Service
 public class AutoService implements IAutoService{
@@ -25,6 +31,13 @@ public class AutoService implements IAutoService{
 
     @Override
     public void deleteAuto(Long id) {
-        autoRepo.deleteById(id);
+        Optional<Auto> optionalAuto = autoRepo.findById(id);
+
+        if (optionalAuto.isPresent()) {
+            autoRepo.deleteById(id);
+        } else {
+            throw new NoSuchElementException("ID de auto no encontrada");
+        }
     }
+
 }
